+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDaoImp.java
@@ -34,13 +34,13 @@ public class UserDaoImp implements UserDao{
     }
 
     @Override
-    public User getUserById(int id) {
+    public User getUserById(Long id) {
         User user = entityManager.find(User.class, id);
         return user;
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(Long id) {
         User userDelete = entityManager.find(User.class, id);
         entityManager.remove(userDelete);
     }
@@ -66,7 +66,7 @@ public class UserDaoImp implements UserDao{
     }
 
     @Override
-    public Role findById(long id) {
+    public Role findById(Long id) {
         return entityManager.find(Role.class, id);
     }
